E3 minmax repaired, alfa beta not, mess otherwise

+++ E3/src/CheckersGridAccessor.java
@@ -1,3 +1,3 @@
 abstract class CheckersGridAccessor {
-    double accessCheckersGrid(CheckersGrid checkersGrid, PlayerColor playerColor, MinMaxEnum whoseTurn) { throw new IllegalStateException("Method not implemented"); }
+    int accessCheckersGrid(CheckersGrid checkersGrid, PlayerColor playerColor, MinMaxEnum whoseTurn) { throw new IllegalStateException("Method not implemented"); }
 }
